Add new documentation files and Docker configuration for Rehau Neasmart 2.0 Gateway
This commit introduces several new files, including `.dockerignore`, `API_DOCUMENTATION.md`, `ARCHITECTURE.md`, and `MIGRATION_GUIDE.md`, enhancing the project's documentation. Additionally, a `docker-compose.yml` file is added to facilitate container orchestration, and the `Dockerfile` is updated for a multi-stage build process. The `README.md` is also updated to reflect these changes, improving the overall usability and deployment readiness of the application.
