+++ env.example
@@ -1,0 +1,54 @@
+# Rehau Neasmart 2.0 Gateway Configuration
+# Copy this file to .env and update with your values
+
+# Server Configuration
+NEASMART_GATEWAY_SERVER_TYPE=tcp
+NEASMART_GATEWAY_SERVER_ADDRESS=192.168.1.100
+NEASMART_GATEWAY_SERVER_PORT=502
+
+# For serial connection (alternative to TCP)
+# NEASMART_GATEWAY_SERVER_TYPE=serial
+# NEASMART_GATEWAY_SERVER_SERIAL_PORT=/dev/ttyUSB0
+# NEASMART_GATEWAY_SERVER_SERIAL_BAUDRATE=38400
+# NEASMART_GATEWAY_SERVER_SERIAL_PARITY=N
+# NEASMART_GATEWAY_SERVER_SERIAL_STOPBITS=1
+# NEASMART_GATEWAY_SERVER_SERIAL_BYTESIZE=8
+
+# API Configuration
+NEASMART_API_HOST=0.0.0.0
+NEASMART_API_PORT=5000
+NEASMART_API_ENABLE_AUTH=false
+NEASMART_API_KEY=your-secure-api-key-here
+NEASMART_API_JWT_SECRET=your-jwt-secret-here
+NEASMART_API_ENABLE_CORS=true
+NEASMART_API_CORS_ORIGINS=*
+NEASMART_API_RATE_LIMIT_PER_MINUTE=60
+
+# Database Configuration
+NEASMART_DATABASE_PATH=./data/registers.db
+NEASMART_DATABASE_TABLE_NAME=holding_registers
+NEASMART_DATABASE_ENABLE_FALLBACK=true
+
+# Modbus Configuration
+NEASMART_MODBUS_SLAVE_ID=240
+NEASMART_MODBUS_SYNC_ON_STARTUP=true
+NEASMART_MODBUS_CIRCUIT_BREAKER_FAILURE_THRESHOLD=5
+NEASMART_MODBUS_CIRCUIT_BREAKER_RECOVERY_TIMEOUT=60
+
+# Logging Configuration
+NEASMART_LOG_LEVEL=INFO
+NEASMART_LOG_FILE=./data/gateway.log
+NEASMART_LOG_ENABLE_FILE=true
+NEASMART_LOG_ENABLE_CONSOLE=true
+NEASMART_LOG_FORMAT=json
+LOG_FORMAT=json
+
+# Production Settings
+ENVIRONMENT=production
+NEASMART_WORKERS=4
+NEASMART_WORKER_CLASS=sync
+NEASMART_WORKER_TIMEOUT=120
+NEASMART_GRACEFUL_TIMEOUT=30
+
+# Debug Mode (set to false in production)
+NEASMART_DEBUG_MODE=false
