+++ install.sh
@@ -90,7 +90,7 @@ echo ""
 echo "Next steps:"
 echo "1. Edit data/config.json with your Modbus device settings"
 echo "2. Run the application:"
-echo "   - Development: python src/main_v2.py"
-echo "   - Production: gunicorn --config src/gunicorn_config.py main_v2:app"
+echo "   - Development: python src/main.py"
+echo "   - Production: gunicorn --config src/gunicorn_config.py main:app"
 echo ""
 echo "For more information, see README.md"
