+++ pytest.ini
@@ -1,0 +1,31 @@
+[pytest]
+# Test discovery patterns
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+
+# Test paths
+testpaths = tests
+addopts = 
+    -v
+    --strict-markers
+    --tb=short
+    --cov=src
+    --cov-report=term-missing
+    --cov-report=html:htmlcov
+    # --cov-fail-under=70  # Uncomment when more tests are added
+
+# Markers for test categorization
+markers =
+    unit: Unit tests (fast, isolated)
+    integration: Integration tests (may require external resources)
+    slow: Slow running tests
+    modbus: Tests requiring Modbus communication
+    database: Tests requiring database
+
+# Logging
+log_cli = true
+log_cli_level = INFO
+
+# Python path
+pythonpath = src
