+++ requirements.txt
@@ -20,3 +20,11 @@ prometheus-client~=0.19.0
 python-json-logger~=2.0.7  # Structured logging
 tenacity~=8.2.3  # Retry library
 cachetools~=5.3.2  # Caching utilities
+
+# Development dependencies (optional)
+pytest~=7.4.3
+pytest-cov~=4.1.0
+pytest-asyncio~=0.21.1
+black~=23.11.0
+flake8~=6.1.0
+mypy~=1.7.1
