+++ src/const.py
@@ -1,7 +1,12 @@
-NEASMART_SYSBUS_BAUD = 38400
+# Modbus communication parameters (from Rehau NEA SMART 2.0 documentation)
+NEASMART_SYSBUS_BAUD = 38400  # Baud rate: 38400 Bits/s
 NEASMART_SYSBUS_DATA_BITS = 8
-NEASMART_SYSBUS_STOP_BITS = 1
-NEASMART_SYSBUS_PARITY = "N"
+NEASMART_SYSBUS_STOP_BITS = 1  # 1 stop bit
+NEASMART_SYSBUS_PARITY = "N"  # None parity
+NEASMART_SLAVE_ADDRESS_PRIMARY = 240  # Primary slave address
+NEASMART_SLAVE_ADDRESS_SECONDARY = 241  # Secondary slave address
+NEASMART_BYTE_ORDER = "MSB"  # Most Significant Byte first
+NEASMART_REGISTER_START = 0  # First address "0"
 NEASMART_BASE_SLAVE_ADDR = 1200
 BASE_ZONE_ID = 100
 ZONE_SETPOINT_ADDR_OFFSET = 1
