--- src/requirements.txt
@@ -1,30 +1,0 @@
-# Core dependencies
-pymodbus~=3.3.2
-Flask~=2.3.2
-sqlitedict~=2.1.0
-
-# Production server (disabled for Python 3.13 compatibility)
-# gunicorn~=21.2.0
-# gevent~=23.9.1
-
-# API enhancements
-flask-cors~=4.0.0
-flask-limiter~=3.5.0
-flask-jwt-extended~=4.5.3
-flasgger~=0.9.7.1  # Swagger UI
-
-# Monitoring and metrics
-prometheus-client~=0.19.0
-
-# Utilities
-python-json-logger~=2.0.7  # Structured logging
-tenacity~=8.2.3  # Retry library
-cachetools~=5.3.2  # Caching utilities
-
-# Development dependencies (optional)
-pytest~=7.4.3
-pytest-cov~=4.1.0
-pytest-asyncio~=0.21.1
-black~=23.11.0
-flake8~=6.1.0
-mypy~=1.7.1
