+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test suite for Rehau Neasmart 2.0 Gateway"""
